pymozhi.py for python2x
How to run the program:=======================
You need utf-8(Unicode) Font like 'karthika.ttf' in your computer to use Pymozhi successfully. You need to type in "Manglish"(phonetic English-Malayalam or anglicized malayalam) to make use of the transliteration scheme, the major strength and distinctive character of this program. You DO NOT NEED the complicated Kerala keyboard map for this program. Down load and keep a printed copy of Manglish for your reference. 